don't test CLI when just testing a subpackage

+++ packages/pymedphys_dicom/tests/test_header_tweaks.py
@@ -46,17 +46,21 @@ ADJUSTED_DICOM_FILENAME = os.path.join(
 
 
 def compare_dicom_cli(command, original, expected):
-    pydicom.write_file(ORIGINAL_DICOM_FILENAME, original)
+    # TODO: Extract CLI definition from subpackages and test separately
 
-    try:
-        subprocess.check_call(command)
-        cli_adjusted_ds = pydicom.read_file(
-            ADJUSTED_DICOM_FILENAME, force=True)
+    if 'SUBPACKAGE' not in os.environ:
 
-        assert str(cli_adjusted_ds) == str(expected)
-    finally:
-        remove_file(ORIGINAL_DICOM_FILENAME)
-        remove_file(ADJUSTED_DICOM_FILENAME)
+        pydicom.write_file(ORIGINAL_DICOM_FILENAME, original)
+
+        try:
+            subprocess.check_call(command)
+            cli_adjusted_ds = pydicom.read_file(
+                ADJUSTED_DICOM_FILENAME, force=True)
+
+            assert str(cli_adjusted_ds) == str(expected)
+        finally:
+            remove_file(ORIGINAL_DICOM_FILENAME)
+            remove_file(ADJUSTED_DICOM_FILENAME)
 
 
 def test_adjust_machine_name():
